Change Quantity to use Double instead of Int
This allows support for fractional quantities for example for shares

+++ Sources/FinanceKit/Quantity.swift
@@ -6,5 +6,5 @@
 
 import Foundation
 
-/// `Quantity` is a typealias for `Int` used to describe the quantity of objects.
-public typealias Quantity = Int
+/// A typealias for `Double` used to describe a fractional (non-whole) quantity.
+public typealias Quantity = Double
